Fix confusion between dto et middlewares
