+++ backend/src/dto/location.dto.js
@@ -1,20 +1,9 @@
-const Joi = require('joi');
-
-const locationSchema = Joi.object({
-  name: Joi.string().max(255).required(),
-  lat: Joi.number().precision(6).optional(),
-  lng: Joi.number().precision(6).optional(),
-});
-
-function validateLocation(req, res, next) {
-  const { error } = locationSchema.validate(req.body, { abortEarly: false });
-  if (error) {
-    return res.status(400).json({
-      message: 'Validation error',
-      details: error.details.map(d => d.message)
-    });
+class LocationDTO {
+  constructor({ name, lat, lng }) {
+    this.name = name;
+    this.lat = lat;
+    this.lng = lng;
   }
-  next();
 }
 
-module.exports = validateLocation;
+module.exports = LocationDTO;
