+++ backend/src/dto/typeDevice.dto.js
@@ -1,18 +1,7 @@
-const Joi = require('joi');
-
-const typeDeviceSchema = Joi.object({
-  name: Joi.string().max(100).required(),
-});
-
-function validateTypeDevice(req, res, next) {
-  const { error } = typeDeviceSchema.validate(req.body, { abortEarly: false });
-  if (error) {
-    return res.status(400).json({
-      message: 'Validation error',
-      details: error.details.map(d => d.message)
-    });
+class TypeDeviceDTO {
+  constructor({ name }) {
+    this.name = name;
   }
-  next();
 }
 
-module.exports = validateTypeDevice;
+module.exports = TypeDeviceDTO;
