+++ backend/src/routes/device.routes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const deviceController = require('../controllers/device.controller');
-const validateDevice = require('../middlewares/deviceValidation');
+const validateDevice = require('../middlewares/device.middleware');
 
 router.get('/', deviceController.getAll.bind(deviceController));
 router.get('/:id', deviceController.getById.bind(deviceController));
